requirements.txt
Flask
pandas
plotly
openpyxl
newspaper3k
requests
selenium
transformers
torch
